Rename ~/rs to ~/routes
