+++ src/components/DisclaimerBar.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom'
 import styled from 'styled-components'
 import { toaster } from 'toasterhea'
 import LightModal from '~/modals/LightModal'
-import { route } from '~/rs'
+import { route } from '~/routes'
 import SvgIcon from '~/shared/components/SvgIcon'
 import { DESKTOP, TABLET } from '~/shared/utils/styled'
 import { Layer } from '~/utils/Layer'
