+++ src/pages/AbstractStreamEditPage/AccessControlSection/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { StreamPermission } from '@streamr/sdk'
 import { useCurrentStreamAbility } from '~/shared/stores/streamAbilities'
 import { StreamDraft } from '~/stores/streamDraft'
-import { route } from '~/rs'
+import { route } from '~/routes'
 import Section from '../Section'
 import { PermissionList } from './PermissionList'
 import { StreamTypeSelector } from './StreamTypeSelector'
