+++ src/pages/ProjectPage/ProjectEditorPage.tsx
@@ -30,7 +30,7 @@ import { Chain } from '~/types'
 import { SalePointsPayload } from '~/types/projects'
 import { formatChainName } from '~/utils'
 import { toastedOperation } from '~/utils/toastedOperation'
-import { route } from '~/rs'
+import { route } from '~/routes'
 import DataUnionFee from './DataUnionFee'
 import DataUnionPayment from './DataUnionPayment'
 import EditorHero from './EditorHero'
