+++ src/stores/streamDraft.tsx
@@ -17,7 +17,7 @@ import { getCurrentChainId } from '~/getters/getCurrentChain'
 import { getStreamrClientInstance } from '~/getters/getStreamrClient'
 import GetCryptoModal from '~/modals/GetCryptoModal'
 import { Bits, ParsedStream, matchBits, parseStream } from '~/parsers/StreamParser'
-import { route } from '~/rs'
+import { route } from '~/routes'
 import InsufficientFundsError from '~/shared/errors/InsufficientFundsError'
 import StreamNotFoundError from '~/shared/errors/StreamNotFoundError'
 import { useCurrentChainId } from '~/shared/stores/chain'
